commit codex 2

+++ audio/hdi_service/primary_impl/src/audio_common.c
@@ -136,7 +136,7 @@ int32_t TimeToAudioTimeStamp(uint64_t bufferFrameSize, struct AudioTimeStamp *ti
 
     time->tvSec += (int64_t)(bufferFrameSize / sampleRate);
 
-    int64_t lastBufFrames = bufferFrameSize % ((int64_t)sampleRate);
+    int64_t lastBufFrames = ((int64_t)bufferFrameSize) % ((int64_t)sampleRate);
 
     time->tvNSec += (lastBufFrames * SEC_TO_NSEC) / ((int64_t)sampleRate);
     if (time->tvNSec >= SEC_TO_NSEC) {
